move: 무한 스크롤 컴포넌트를 분리
